app/requirements.txt
streamlit
pandas
mysql-connector-python
requests
elasticsearch
python-dotenv
matplotlib
scikit-learn  
numpy         
